+++ src/test/java/io/outfoxx/typescriptpoet/test/InterfaceSpecTests.kt
@@ -118,7 +118,11 @@ class InterfaceSpecTests {
       out.toString(),
       emits(
         """
-            interface Test<X extends Test2, Y extends Test3 & Test4, Z extends Test5 | keyof Test6> {}
+            interface Test<
+              X extends Test2,
+              Y extends Test3 & Test4,
+              Z extends Test5 | keyof Test6,
+            > {}
 
         """.trimIndent(),
       ),